pyproject.toml
[tool.poetry]
name = "minimal-update-server"
version = "0.0.1"
description = "Minimal ownCloud update server implementation"
authors = ["Fabian Müller <fmueller@owncloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.1.2"
PyYAML = "^6.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
